Exception handled

+++ src/main/java/in/dmart/oms/models/Customer.java
@@ -7,7 +7,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name="customer")
+@Table(name = "customer")
 public class Customer {
 
 	@Id
@@ -15,47 +15,59 @@ public class Customer {
 	private int customer_id;
 	private String name;
 	private String email;
-	private int phone;
+	private long phone;
 	private String address;
 	private String location;
-	
+
 	public int getCustomer_id() {
 		return customer_id;
 	}
+
 	public void setCustomer_id(int customer_id) {
 		this.customer_id = customer_id;
 	}
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public String getEmail() {
 		return email;
 	}
+
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public int getPhone() {
+
+	public long getPhone() {
 		return phone;
 	}
-	public void setPhone(int phone) {
+
+	public void setPhone(long phone) {
 		this.phone = phone;
 	}
+
 	public String getAddress() {
 		return address;
 	}
+
 	public void setAddress(String address) {
 		this.address = address;
 	}
+
 	public String getLocation() {
 		return location;
 	}
+
 	public void setLocation(String location) {
 		this.location = location;
 	}
-	public Customer(int customer_id, String name, String email, int phone, String address, String location) {
+
+	public Customer(int customer_id, String name, String email, long phone, String address, String location) {
 		super();
 		this.customer_id = customer_id;
 		this.name = name;
@@ -64,15 +76,16 @@ public class Customer {
 		this.address = address;
 		this.location = location;
 	}
+
 	public Customer() {
 		super();
-		
+
 	}
+
 	@Override
 	public String toString() {
 		return "Customer [customer_id=" + customer_id + ", name=" + name + ", email=" + email + ", phone=" + phone
 				+ ", address=" + address + ", location=" + location + "]";
 	}
-	
-	
+
 }
